Added a new version of the sampling game
I changed the method of executing the game via a class (called game) and a main loop function which prompts the user to play again or quit the game. In addition, I added an empty list to record over- or underestimates as the user continues to play. Now, I will look at making a GUI which includes a figure (histogram, centered on zero) with a running update as the user plays the game.