Update sampling_game.py
I changed the code to be object oriented and I added several lists to store the data as the game is played. I also added a plotting function to visualize the guesses after the game is completed. This might be better implemented as a figure that updates while the game is being played. I still need to address error handling processes throughout the code.