Added histogram to GUI and improved button states
The histogram was added but it requires some work. It currently plots over itself after every game and it needs axis titles etc. Button states were improved so that the user could not have a second attempt after the answer was revealed. Also the return key has been bound to the entry field as a second option to the submit button.